Keen.
Never heard of it. 

In Jessamine County we have a Kenne. 

I can’t help but to think this is another display of their pride. 

It’s wrecked. Unknown brand. Purple. And just where I happen to be. 

Creator just confirmed. It’s pride on display. I do not know why the fuck this Company exists. And up pulls Alan Houps lookalike. 

He was here to distract me from the hit. 

My home of Jessamine County has to be destroyed. They’ve killed so many.